requirements.txt
Django>=4.0
djangorestframework>=3.14.0
psutil>=5.9.0
requests>=2.25.0